fix: refactor tab type and settings constants into a new settings-types.ts file
