+++ apps/frontend/src/components/ide-shared/GitLabImportDialog.tsx
@@ -293,9 +293,7 @@ export function GitLabImportDialog({
 
       // Detect characters from imported RPY files
       try {
-        // Fast skip: don't detect characters if this import was superseded
-        if (currentImportId !== importIdRef.current) return;
-
+        // react-doctor-disable-next-line react-doctor/async-defer-await
         const detectionResult = await charactersApi.detectCharacters(
           result.project.id
         );
