+++ apps/frontend/src/components/ide-shared/ZipImportFilesDialog.tsx
@@ -253,13 +253,9 @@ export function ZipImportFilesDialog({
           console.error("Failed to refresh cache after import:", cacheError);
         }
 
-        if (currentImportId !== importIdRef.current) return;
-
         // Detect characters from imported RPY files
         try {
-          // Fast skip: don't detect characters if this import was superseded
-          if (currentImportId !== importIdRef.current) return;
-
+          // react-doctor-disable-next-line react-doctor/async-defer-await
           const detectionResult =
             await charactersApi.detectCharacters(projectId);
