+++ apps/frontend/src/components/ide-shared/ZipImportProjectDialog.tsx
@@ -240,9 +240,7 @@ export function ZipImportProjectDialog({
 
       // Detect characters from imported RPY files
       try {
-        // Fast skip: don't detect characters if this import was superseded
-        if (currentImportId !== importIdRef.current) return;
-
+        // react-doctor-disable-next-line react-doctor/async-defer-await
         const detectionResult = await charactersApi.detectCharacters(
           data.project.id
         );
