+++ apps/frontend/src/components/ide-shared/settings-types.ts
@@ -1,0 +1,8 @@
+export type Tab = "user" | "projects" | "integrations" | "system";
+
+export const SETTINGS_TABS = [
+  "user",
+  "projects",
+  "integrations",
+  "system",
+] as const;
